code/requirements.txt
asgiref==3.7.2
Django==4.2.11
sqlparse==0.4.4
psycopg2-binary==2.9.9
django-crispy-forms==2.1.0
crispy-bootstrap5==2024.2
django-allauth==0.61.1
#django-oauth-toolkit==2.3.0
environs[django]==9.5.0
whitenoise==6.6.0
gunicorn==21.2.0
djangorestframework==3.15.0
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.3.4